+++ src/BayesFilters/include/BayesFilters/LikelihoodModel.h
@@ -15,7 +15,7 @@ class bfl::LikelihoodModel
 public:
     virtual ~LikelihoodModel() noexcept { };
 
-    virtual std::pair<bool, Eigen::VectorXf> likelihood(const MeasurementModel& measurement_model, const Eigen::Ref<const Eigen::MatrixXf>& pred_states) = 0;
+    virtual std::pair<bool, Eigen::VectorXd> likelihood(const MeasurementModel& measurement_model, const Eigen::Ref<const Eigen::MatrixXd>& pred_states) = 0;
 };
 
 #endif /* LIKELIHOODMODEL_H */
